fix(tests): correggi 39 test failures in Cespiti/Erogazioni/Relazione/F24/Policies
Correzioni ai modelli:
- Member: aggiunge scopeAttivi() usato da RelazioneMissioneService
- CodiceIva: aggiunge default $attributes per di_sistema=false, attivo=true
- LiquidazioneIva: aggiunge default $attributes per status='bozza'
- Asset: aggiunge HasFactory trait + AssetFactory
- CooperativeSharePolicy: corregge $share->stato → $share->status

Correzioni ai service:
- AmortizzamentoService: cast (float) su max(0,...) per valore_residuo_fine_anno
- CespitiService: cast (float) su residuo + fix confronto date Carbon vs string in aggiorna()

Correzioni alle route:
- web.php: aggiunge role middleware a f24.index (e compensi-terzi.index già aggiunto)

Correzioni ai test (pattern standard):
- GET Inertia tests: X-Inertia header + JSON parse invece di withoutMiddleware+assertInertia
- POST/PUT/DELETE tests: aggiunge withoutMiddleware([ValidateCsrfToken, AuthenticateSession])
- PoliciesTest: makeShare() crea membro reale + usa LiquidazioneIva reale per FK
- MemberTypeSeeder: guard app()->bound() prima di app('current_tenant') per evitare crash

+++ app/Services/AmortizzamentoService.php
@@ -132,7 +132,7 @@ class AmortizzamentoService
         }
 
         $fondoFine   = round($fondoInizio + $quotaCalcolata, 2);
-        $valoreResiduo = max(0, round($costoStorico - $fondoFine, 2));
+        $valoreResiduo = (float) max(0, round($costoStorico - $fondoFine, 2));
 
         return [
             'quota_calcolata'          => $quotaCalcolata,
